class-based split for meta-learning
This will hopefully prove useful for meta-learning.

The hill-climbing algorithm used in this program is extremely useful. With a random class split, we'd have to sacrifice upwards of 3K samples (out of 22K). With hill-climbing, this number quickly shrinks to ~400 samples lost. In other words, we go from losing 14% of samples to 2% of them.